fix: update autowiring to support spring 4.2.0
